pyproject.toml
[tool.poetry]
name = "backend-services-reference"
version = "1.0.0"
description = ""
authors = ["Juan Gómez <juan@wordbox.ai>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-logging = "^1.15.1"
neo4j = "1.7.6"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
coverage = "^5.2.1"
pre-commit = "^2.7.1"
flake8 = "^3.8.3"
mypy = "^0.782"
bandit = "*"

[tool.poetry.scripts]
mypy_task = "scripts:mypy_task"
bp_tests = "scripts:bp_tests"
endpoints_tests = "scripts:endpoint_tests"
coverage_task = "scripts:coverage_task"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
